refactor(frontend): move auth token controls to a separate directory
The auth token controls start to become more complex and it is wise to
start extracting them into multiple files in a separate directory.
