+++ frontend/src/auth/controls/ChangeAuthenticationTokenButton.tsx
